// examples/integrate-gql-tada/graphql-env.d.ts
export type introspection = {
  name: never
  query: 'Query'
  mutation: 'Mutation'
  subscription: never
  types: introspection_types
}